Separated classes (and 2 functions not belonging to a class) into separate files
